Added parts of the database, renamed some files and also changed the whole app creation system
