+++ app/models/user.py
@@ -1,0 +1,23 @@
+from app.__init__ import db
+from datetime import datetime
+
+class User(db.Models):
+    id = db.Column(db.Integer, primary_key=True)
+    firstname = db.Column(db.String(64), index=True, unique=True)
+    lastname = db.Column(db.String(64), index=True)
+    username = db.Column(db.String(64), index=True)
+    email = db.Column(db.String(120), index=True, unique=True)
+    phone = db.Column(db.Integer, index=True, unique=True)
+    password_hash = db.Column(db.String(128))
+    
+    def __repr__(self):
+        return '<User {}>'.format(self.username)
+    
+class Post(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    body = db.Column(db.String(140))
+    timestamp = db.Column(db.DateTime, index=True, default=datetime.utcnow)
+    user_id = db.Column(db.Integer, db.ForeignKey('user.id'))
+    
+    def __repr__(self):
+        return '<Post {}>'.format(self.body)
